+++ src/Controller/UserController.php
@@ -3,8 +3,8 @@
 namespace App\Controller;
 
 use App\Entity\User;
-use App\Form\PasswordResetType;
-use App\Form\UserType;
+use App\Form\Type\PasswordResetType;
+use App\Form\Type\UserType;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\HttpFoundation\Request;
