+++ src/Form/Type/PasswordResetType.php
@@ -1,18 +1,16 @@
 <?php
 
-namespace App\Form;
+namespace App\Form\Type;
 
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\PasswordType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
-use Symfony\Component\Form\Extension\Core\Type\TextType;
-use Symfony\Component\Form\Extension\Core\Type\EmailType;
 use Symfony\Component\Form\Extension\Core\Type\RepeatedType;
 use Symfony\Component\Form\Extension\Core\Type\SubmitType;
-use Gregwar\CaptchaBundle\Type\CaptchaType;
 
-class UserType extends AbstractType
+
+class PasswordResetType extends AbstractType
 {
     /**
      * {@inheritdoc}
@@ -20,12 +18,6 @@ class UserType extends AbstractType
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
         $builder
-            ->add('username', TextType::class, array(
-                'label' => false,
-                'attr' => array(
-                    'placeholder' => 'Choose a username'
-                )
-            ))
             ->add('password',RepeatedType::class, array(
                 'type' => PasswordType::class,
                 'invalid_message' => 'The password are not the same !',
@@ -40,30 +32,6 @@ class UserType extends AbstractType
                     'attr' => array(
                         'placeholder' => 'Repeat your Password'
                     )
-                )
-            ))
-            ->add('email', RepeatedType::class, array(
-                'type' => EmailType::class,
-                'invalid_message' => 'The emails are not the same !',
-                'first_options' => array(
-                    'label' => false,
-                    'attr' => array(
-                        'placeholder' => 'Your email'
-                    )
-                ),
-                'second_options' => array(
-                    'label' => false,
-                    'attr' => array(
-                        'placeholder' => 'Repeat you email'
-                    )
-                )
-            ))
-            ->add('captcha', CaptchaType::class, array(
-                'label' => false,
-                'reload' => true,
-                'as_url' => true,
-                'attr' => array(
-                    'placeholder' => 'Enter the same text as above'
                 )
             ))
             ->add('submit', SubmitType::class, array(
