+++ src/Form/Type/UserType.php
@@ -1,16 +1,18 @@
 <?php
 
-namespace App\Form;
+namespace App\Form\Type;
 
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\PasswordType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
+use Symfony\Component\Form\Extension\Core\Type\TextType;
+use Symfony\Component\Form\Extension\Core\Type\EmailType;
 use Symfony\Component\Form\Extension\Core\Type\RepeatedType;
 use Symfony\Component\Form\Extension\Core\Type\SubmitType;
+use Gregwar\CaptchaBundle\Type\CaptchaType;
 
-
-class PasswordResetType extends AbstractType
+class UserType extends AbstractType
 {
     /**
      * {@inheritdoc}
@@ -18,6 +20,12 @@ class PasswordResetType extends AbstractType
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
         $builder
+            ->add('username', TextType::class, array(
+                'label' => false,
+                'attr' => array(
+                    'placeholder' => 'Choose a username'
+                )
+            ))
             ->add('password',RepeatedType::class, array(
                 'type' => PasswordType::class,
                 'invalid_message' => 'The password are not the same !',
@@ -32,6 +40,30 @@ class PasswordResetType extends AbstractType
                     'attr' => array(
                         'placeholder' => 'Repeat your Password'
                     )
+                )
+            ))
+            ->add('email', RepeatedType::class, array(
+                'type' => EmailType::class,
+                'invalid_message' => 'The emails are not the same !',
+                'first_options' => array(
+                    'label' => false,
+                    'attr' => array(
+                        'placeholder' => 'Your email'
+                    )
+                ),
+                'second_options' => array(
+                    'label' => false,
+                    'attr' => array(
+                        'placeholder' => 'Repeat you email'
+                    )
+                )
+            ))
+            ->add('captcha', CaptchaType::class, array(
+                'label' => false,
+                'reload' => true,
+                'as_url' => true,
+                'attr' => array(
+                    'placeholder' => 'Enter the same text as above'
                 )
             ))
             ->add('submit', SubmitType::class, array(
